requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain-anthropic
python-decouple
uvicorn[standard]
